refactor: Repository 인터페이스화
