+++ src/main/java/racingcar/repository/RacingCarRepository.java
@@ -1,0 +1,12 @@
+package racingcar.repository;
+
+import java.util.List;
+
+import racingcar.domain.Cars;
+import racingcar.dto.ResultDto;
+
+public interface RacingCarRepository {
+	void save(Cars cars, int count);
+
+	List<ResultDto> find();
+}
