+++ src/main/java/racingcar/repository/console/ConsoleRacingCarRepository.java
@@ -1,0 +1,18 @@
+package racingcar.repository.console;
+
+import java.util.List;
+
+import racingcar.domain.Cars;
+import racingcar.dto.ResultDto;
+import racingcar.repository.RacingCarRepository;
+
+public class ConsoleRacingCarRepository implements RacingCarRepository {
+	@Override
+	public void save(Cars cars, int count) {
+	}
+
+	@Override
+	public List<ResultDto> find() {
+		return null;
+	}
+}
